cleared casting warnings
git-svn-id: svn://orca.physics.unc.edu/Orca64@11953 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/HV/UnivVoltHV/ORCircularBufferUV.m
@@ -37,7 +37,7 @@ NSString* CBeValue		= @"Value";
 }
 
 //mah -- changed to count to remove compiler warning with other methods named 'size' in other objects that return diff type than int32_t.
-- ( int32_t ) count
+- ( NSUInteger ) count
 {
 	return (uint32_t)[mStorageArray count];
 }
